--- cloudbuild.yaml
@@ -1,29 +1,0 @@
-steps:
-  - name: 'gcr.io/cloud-builders/docker'
-    args:
-      - 'build' 
-      - '-t' 
-      - 'gcr.io/$PROJECT_ID/scrape'
-      - '--build-arg'
-      - 'projectid=$PROJECT_ID'
-      - '.'
-
-  - name: 'gcr.io/cloud-builders/docker'
-    args: 
-      - 'push'
-      - 'gcr.io/$PROJECT_ID/scrape'
-
-  - name: 'gcr.io/cloud-builders/gcloud'
-    args:
-      - 'run'
-      - 'jobs'
-      - 'deploy'
-      - 'scrape'
-      - '--image'
-      - 'gcr.io/$PROJECT_ID/scrape'
-      - '--region'
-      - 'asia-northeast1'
-      - '--max-retries'
-      - '0'
-      - '--memory'
-      - '600Mi'
