+++ kabumap-scrape/cloudbuild.yaml
@@ -3,7 +3,7 @@ steps:
     args:
       - 'build' 
       - '-t' 
-      - 'asia-northeast1-docker.pkg.dev/$PROJECT_ID/abc/test'
+      - 'asia-northeast1-docker.pkg.dev/$PROJECT_ID/abc/scrape'
       - '--build-arg'
       - 'projectid=$PROJECT_ID'
       - '.'
@@ -11,18 +11,20 @@ steps:
   - name: 'gcr.io/cloud-builders/docker'
     args: 
       - 'push'
-      - 'asia-northeast1-docker.pkg.dev/$PROJECT_ID/abc/test'
+      - 'asia-northeast1-docker.pkg.dev/$PROJECT_ID/abc/scrape'
 
-  - name: 'gcr.io/cloud-builders/gcloud'
+  - name: 'gcr.io/google.com/cloudsdktool/cloud-sdk:slim'
     args:
+      - 'gcloud'
       - 'run'
       - 'jobs'
       - 'deploy'
-      - 'test'
+      - 'scrape'
       - '--image'
-      - 'asia-northeast1-docker.pkg.dev/$PROJECT_ID/abc/test'
+      - 'asia-northeast1-docker.pkg.dev/$PROJECT_ID/abc/scrape'
       - '--region'
       - 'asia-northeast1'
       - '--max-retries'
       - '0'
-      - '--allow-unauthenticated'
+      - '--memory'
+      - '800Mi'
