+++ kabumap-scrape/requirements.txt
@@ -1,0 +1,6 @@
+pandas
+numpy
+playwright
+google-cloud-storage
+google-cloud-bigquery
+jpholiday
