--- readme.md
@@ -1,1 +1,0 @@
-## This is my git practice
