--- requirements.txt
@@ -1,4 +1,0 @@
-pandas
-playwright
-google-cloud-storage
-google-cloud-bigquery
