--- scheduler/Dockerfile
@@ -1,9 +1,0 @@
-FROM python:3.9-slim
-
-WORKDIR /app
-
-COPY . .
-
-CMD ["python", "test.py"]
-
-
