--- scheduler/test.py
@@ -1,1 +1,0 @@
-print("Hi this is test script")
